backend/requirements.txt
fastapi
uvicorn
boto3==1.34.34
google-cloud-resource-manager==1.12.3
google-cloud-billing==1.12.1
google-cloud-bigquery==3.13.0
google-cloud-storage==2.10.0
google-auth==2.23.4
pydantic==2.5.3